rename to index.js and update snapshots etc accordingly
