+++ packages/package-json/test/fixtures/testPackageJsonNoScripts.json
